[src] KDE-based moment frequency prior
Considering the interest on only-tef by itself, it was worth to
implement another baseline that gauge the relevance of the moments by
itself. We rely on KDE to build a 2D-PDF function of start and duration
of the moments based on the training set. During testing, we score any
proposal based on the PDF.

- The main advantage of modeling the 2D-PDF is that we can plug-in the
  duration of the video as a normalization factor of the segments
  without any problem. Considering the length of the videos could be
  relevant as people don't record moments randomly. Similarly,
  annotators come-up with description based on the whole picture.

- There is nothing wrong with our discrte approach. We could probably
  have to tweak it more. However, it was difficult to incorporate the
  length of the videos there.

- We did not tweak the bandwidth or kernel of the KDE algorithm. We
  rather rely of the ad-hoc method provided by scipy. This could be
  helpful to make the baseline better, but we don't have time to do it.