[scripts] Update single video retrieval evaluation
Due to a mismatch btw the original python metrics and our pytorch
version, we have to refactor, add a bit of verbosity and make this
program even more convoluted. After a while, we realize that torch is
not as programmer friendly as python and we forgot to pass `topk` as
a float tensor. This introduced a systematic error due to casting the
average ranks.

- Add documentation to `didemo_evaluation` to at least notifiy that
  we must use floats and reduce debugging time.

- Minor modification to `rank` to don't force tuple if not needed. This
  was introduced to make our program less verbose, and avoid copying of
  data.

- Add documentation to the program and make explicit note that we should
  clean it.

- Refactor the code adding a function to load weights.

- Add or rename optional arguments. In particular:
  (i) rename `old` with its purpose, selecting the module required for
      the Dataset abstraction.
  (ii) add `python-eval` to use the original functions working with
       python data types.