[src] Alternative when cannot find negative sample
Given that the proposal sampling scheme relies on the annotations and
the proposals, it's possible that a given configuration won't return
negative proposals that satisfy the criteria. In such cases, we can
resort in the random proposal generation. This turned out to be the case
in ActivityNet with the hyper-parameter config explored before.