[src] Intra-negative sampling next to moments
Sample negative proposals inside the video right next to the moment. The
argument is given as a probability to biased the sampling of moment with
such characteristics. We can disable by setting the probability to be a
negative number. In case, the proposal can't no be fitted either to the
right or left, it falls back into generating a proposal with random
start and end inside the video.