[WIP] Update corpus_retrieval_2nd_eval program
The program isn't ready to run experiments, as we identify that it
doesn't produce consistent results accross multiple runs. This commit
only makes tje program functional again by adding the `setup_snapshot_`
to the end of the file, after removing it from `utils` in the previous
commit. Add few comments regarding how to handle snapshots, and remove
the `try` statement as it hides the errors.